+++ apps/E2E/wdio.conf.ios.js
@@ -14,8 +14,9 @@ exports.config = {
     {
       maxInstances: 1, // Maximum number of total parallel running workers.
       platformName: 'iOS',
+      // Keep this in sync with the simulator we run in Azure Pipelines
       'appium:platformVersion': '16.2',
-      'appium:deviceName': 'iPhone 13',
+      'appium:deviceName': 'iPhone 14',
       'appium:automationName': 'XCUITest',
       'appium:bundleId': 'com.microsoft.ReactTestApp',
     },