+++ packages/codemods/src/transform.ts
@@ -28,7 +28,7 @@ export const yargsParse = (args: string[]): argsType => {
       choices: ['button-v0-to-v1', 'deprecate-exports'],
     })
     .demandOption(['path', 'transform'])
-    .parse(args);
+    .parseSync(args);
 };
 
 export const transform = (args: argsType) => {