+++ packages/components/Button/src/FAB/FAB.tsx
@@ -1,3 +1,4 @@
+/** @jsxRuntime classic */
 /** @jsx withSlots */
 import type * as React from 'react';
 import { Pressable } from 'react-native';