+++ packages/theming/apple-theme/src/createAppleTheme.macos.ts
@@ -19,6 +19,8 @@ export function createAppleTheme(): ThemeReference {
   // Fetch initial system settings for high contrast mode
   highContrastHandler();
   // Invalidate theme and set prop when high contrast setting changes
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
   AccessibilityInfo.addEventListener('highContrastChanged', () => {
     highContrastHandler();
   });
@@ -30,6 +32,8 @@ export function createAppleTheme(): ThemeReference {
 }
 
 function highContrastHandler() {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
   AccessibilityInfo.isHighContrastEnabled().then((isEnabled) => {
     setIsHighContrast(isEnabled);
     appleThemeReference.invalidate();