+++ packages/deprecated/foundation-tokens/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log - @uifabricshared/foundation-tokens
 
-This log was last generated on Tue, 30 May 2023 20:34:15 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jun 2023 19:26:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.0
+
+Mon, 05 Jun 2023 19:26:25 GMT
+
+### Minor changes
+
+- Upgrade to React Native 0.71 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/merge-props to v0.6.0
+- Bump @fluentui-react-native/tokens to v0.21.0
+- Bump @uifabricshared/foundation-settings to v0.13.0
+- Bump @fluentui-react-native/memo-cache to v1.2.0
 
 ## 0.12.33