+++ packages/experimental/Spinner/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui-react-native/spinner
 
-This log was last generated on Tue, 30 May 2023 20:34:15 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jun 2023 19:26:24 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.0
+
+Mon, 05 Jun 2023 19:26:24 GMT
+
+### Minor changes
+
+- Upgrade to React Native 0.71 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/component-cache to v1.5.0
+- Bump @fluentui-react-native/theme-tokens to v0.25.0
+- Bump @fluentui-react-native/use-styling to v0.10.0
+- Bump @fluentui-react-native/framework to v0.10.0
+- Bump @fluentui-react-native/text to v0.20.0
 
 ## 0.4.2