+++ packages/framework/themed-stylesheet/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/themed-stylesheet
 
-This log was last generated on Fri, 24 Feb 2023 20:40:00 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jun 2023 19:26:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.5.0
+
+Mon, 05 Jun 2023 19:26:25 GMT
+
+### Minor changes
+
+- Upgrade to React Native 0.71 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/memo-cache to v1.2.0
 
 ## 1.4.3