+++ packages/theming/theme-tokens/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/theme-tokens
 
-This log was last generated on Tue, 30 May 2023 20:34:15 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jun 2023 19:26:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.25.0
+
+Mon, 05 Jun 2023 19:26:25 GMT
+
+### Minor changes
+
+- Upgrade to React Native 0.71 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/theme-types to v0.32.0
 
 ## 0.24.7