+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,22 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Tue, 30 May 2023 20:34:15 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Jun 2023 19:26:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.26.0
+
+Mon, 05 Jun 2023 19:26:25 GMT
+
+### Minor changes
+
+- Upgrade to React Native 0.71 (sanajmi@microsoft.com)
+- Bump @fluentui-react-native/default-theme to v0.19.0
+- Bump @fluentui-react-native/memo-cache to v1.2.0
+- Bump @fluentui-react-native/theme-tokens to v0.25.0
+- Bump @fluentui-react-native/theme-types to v0.32.0
+- Bump @fluentui-react-native/theme to v0.9.0
+- Bump @fluentui-react-native/theming-utils to v0.24.0
 
 ## 0.25.12